requirements.txt
requests
python-dotenv
pyspark
apache-airflow
snowflake-connector-python
pandas